+++ src/main/java/com/prgrms/management/customer/domain/CustomerType.java
@@ -1,5 +1,7 @@
 package com.prgrms.management.customer.domain;
 
+import com.prgrms.management.command.domain.Command;
+
 import java.util.Arrays;
 import java.util.NoSuchElementException;
 
@@ -10,6 +12,6 @@ public enum CustomerType {
         return Arrays.stream(CustomerType.values())
                 .filter(e -> e.name().equals(input.toUpperCase()))
                 .findAny()
-                .orElseThrow(()-> new NoSuchElementException("존재하지 않는 손님 타입입니다."));
+                .orElseThrow(()-> new NoSuchElementException(CustomerType.class+":존재하지 않는 손님 타입입니다."));
     }
 }
