+++ src/main/java/com/prgrms/management/customer/repository/CustomerRepository.java
@@ -1,12 +1,11 @@
 package com.prgrms.management.customer.repository;
 
 import com.prgrms.management.customer.domain.Customer;
-import com.prgrms.management.customer.domain.CustomerType;
 
 import java.util.List;
-import java.util.UUID;
 
 public interface CustomerRepository {
     Customer insert(Customer customer);
+
     List<Customer> findBlackList();
 }
