+++ src/main/resources/customer_blacklist.csv
@@ -3,5 +3,4 @@
 7480d7fa-abc2-488e-89a3-9958f8a3cd2d,NORMAL
 92f41ccf-0348-4671-b4ca-8941b2130a18,NORMAL
 2b9a4c5d-6e68-47b2-8620-83625e0b7efe,NORMAL
-e8495079-f21c-4ea3-9aa4-a3d516b864b9,NORMAL48fc050f-0a41-4eca-bb5a-82fed57e3eca,NORMAL
-e6cad910-24da-4913-8b5e-0417f4ec70fd,NORMAL
+461af779-ee41-48e2-bf8a-ce232d065e59,NORMAL
