+++ src/main/resources/logFile.log
@@ -1,3 +1,5 @@
-21:44:01.360 [main] WARN  c.prgrms.management.command.Console - class java.util.NoSuchElementException:없는 명령어입니다.
-21:44:02.903 [main] WARN  c.prgrms.management.command.Console - class java.util.NoSuchElementException:없는 명령어입니다.
-21:44:04.282 [main] WARN  c.prgrms.management.command.Console - class java.util.NoSuchElementException:없는 명령어입니다.
+23:54:59.756 [main] WARN c.p.m.command.CommandService - class java.util.NoSuchElementException:class com.prgrms.management.command.domain.Command:없는 명령어입니다.
+23:55:23.295 [main] WARN c.p.m.command.CommandService - class java.util.NoSuchElementException:class com.prgrms.management.command.domain.Command:없는 명령어입니다.
+23:55:28.211 [main] WARN c.p.m.command.CommandService - class java.util.NoSuchElementException:class com.prgrms.management.voucher.domain.VoucherType:잘못된 입력 값입니다.
+23:55:39.745 [main] WARN c.p.m.command.CommandService - class java.lang.NumberFormatException:class com.prgrms.management.voucher.domain.VoucherType:0~10000 이내로 입력하세요
+23:55:53.232 [main] WARN c.p.m.command.CommandService - class java.lang.NumberFormatException:class com.prgrms.management.voucher.domain.VoucherType:0~100 이내로 입력하세요
