+++ src/main/resources/voucher.csv
@@ -11,3 +11,7 @@ d0b144ea-1082-4228-8f20-685ee0a51192,100,FIXED
 c2533acf-ca23-4a9b-8bf7-b80dd6b63178,5,PERCENT
 efdc1d06-617c-4fcb-b7a5-f5eef6b1dabb,100,FIXED
 bf24a46b-e8b8-4abb-8782-f3e38066ac3c,5,PERCENT
+e2e7b6f9-891d-43e0-846c-85fadd96248c,100,FIXED
+4354962d-7d9e-4a55-b3f0-20a624bbc4a6,5,PERCENT
+866932a2-6db2-4e7b-82ea-b708afbf3449,100,FIXED
+fcfe733b-e365-43c7-9e78-17e4342e5253,5,PERCENT
